@@ -11,19 +11,22 @@
     },
     "devDependencies": {
         "axios": "^0.19",
-        "bootstrap": "^4.0.0",
+        "bootstrap": "^4.4.1",
         "cross-env": "^7.0",
-        "jquery": "^3.2",
+        "jquery": "^3.4.1",
         "laravel-mix": "^5.0.1",
         "lodash": "^4.17.13",
-        "popper.js": "^1.12",
-        "resolve-url-loader": "^2.3.1",
-        "sass": "^1.20.1",
+        "popper.js": "^1.16.1",
+        "resolve-url-loader": "^3.1.0",
+        "sass": "^1.15.2",
         "sass-loader": "^8.0.0",
-        "vue": "^2.5.17",
-        "vue-template-compiler": "^2.6.10"
+        "vue": "^2.6.11",
+        "vue-template-compiler": "^2.6.11",
+        "vuetify": "^2.2.19",
+        "vuex": "^3.1.3"
     },
-    "dependencies": {
-        "vuetify": "^2.2.18"
+    "dependencies": {},
+    "config": {
+        "webpack": "node_modules/laravel-mix/setup/webpack.config.js"
     }
 }
