@@ -1,19 +1,18 @@
 <?php
 
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Route;
 
-/*
-|--------------------------------------------------------------------------
-| API Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register API routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| is assigned the "api" middleware group. Enjoy building your API!
-|
-*/
 
-Route::middleware('auth:api')->get('/user', function (Request $request) {
+ Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
-});
+ });
+
+Route::post('/', 'SsoGroupProfileTypeController@store');
+
+Route::get('/', 'SsoGroupProfileTypeController@get');
+
+Route::get('/{filter}/{filterState}', 'SsoGroupProfileTypeController@search');
+
+Route::match(['put'], '/{id}', 'SsoGroupProfileTypeController@edit');
+
+Route::delete('/{id}', 'SsoGroupProfileTypeController@delete');
