@@ -1,7 +1,5 @@
 <?php
 
-use Illuminate\Support\Facades\Route;
-
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -13,10 +11,8 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('/', function () {
-    return view('welcome');
-});
-
+//
 Auth::routes();
 
-Route::get('/home', 'HomeController@index')->name('home');
+Route::get('/', "SsoGroupProfileTypeController@index");
+Route::get('/home', "SsoGroupProfileTypeController@index");
